requirements.txt
asgi-redis==1.4.1
asgiref==1.1.2
attrs==17.4.0
autobahn==18.4.1
Automat==0.6.0
channels==1.1.5
constantly==15.1.0
daphne==1.4.2
diff-match-patch==20121119
Django==1.10.7
django-background-tasks==1.1.13
django-compat==1.0.15
django-filter==1.1.0
django-import-export==0.4.5
djangorestframework==3.8.2
et-xmlfile==1.0.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
jdcal==1.4
Markdown==2.6.11
msgpack-python==0.5.6
odfpy==1.3.6
openpyxl==2.5.2
psycopg2==2.7.4
pypiwin32==223
pywin32==223
PyYAML==3.12
redis==2.10.6
six==1.11.0
tablib==0.12.1
Twisted==17.9.0
txaio==2.10.0
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.4.3
